+++ sqlstore/src/main/resources/db/migration/V8510__timepoint.sql
@@ -1,0 +1,1 @@
+ALTER TABLE Sample ADD COLUMN timepoint varchar(50);